Add slider module
